Add OVN schema version checks in Makefile.am
This was left out during the OVS/OVN split.

We still need to handle the below in the automake files
 - OVN northbound/southbound E-R diagram
 - OVN northbound/southbound schema documentation
 - ovn-architecture manpage generation.

The above requires some amount of work (mainly in pointing to the correct
path for ovsdb-doc/ovsdb-dot.in) and will be added in a separate patch.

OVN schema version checks is very important and hence only that is considered
in this patch.

Signed-off-by: Numan Siddique <nusiddiq@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>