vagrant: Port Vagrantfile to build and test OVN
Revisit Vagrantfile for OVN, so these distros can be used
to build OVN and run its tests:
  - centos7
  - debian-8
  - fedora-29

Vagrant box fedora/23-cloud-base is not working anymore, so it
got bumped to fedora-29.

This patch assumes that OVS source is located on same directory
where this OVN source is located (i.e.: ../ovs).

Provisioning makes no attempt to build and install openvswitch
kernel modules, but it will build ovs on each VM in order to
support OVN.

Signed-off-by: Flavio Fernandes <flavio@flaviof.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>