tests/system-ovn: Ignore some ovn-controller warnings
This log message was introduced in commit 5344f24ecb.  It may be more
appropriate as a deubg message, but as a warning, it breaks this test suite.
Filtering it out of the logs gets these tests passing for me.

A sample of the messages encountered in a test run are:

2019-10-25T01:06:53.026Z|00010|chassis|WARN|Could not find Chassis : stored (hv1) ovs (hv1)
2019-10-25T01:06:53.026Z|00011|chassis|WARN|Could not find Chassis : stored (hv1) ovs (hv1)
2019-10-25T01:06:53.026Z|00013|chassis|WARN|Could not find Chassis : stored (hv1) ovs (hv1)
2019-10-25T01:06:53.026Z|00014|chassis|WARN|Could not find Chassis : stored (hv1) ovs (hv1)

Signed-off-by: 0-day Robot <robot@bytheb.org>