vagrant: Use python3 and newer linux distros
Stop usage of python2 as ovs+ovn no longer support it.
Update vagrant boxes to the following revisions:

- Debian: buster (from jessie)
- Fedora: v31 (from v29)
- Centos: 8 (from 7, kinda**)

Centos 7 may still be used, but only if explicitly
provided in command: 'vagrant up centos-7'

Fedora-31's dnf is not reliable. This patch
uses a max retry loop to improve the odds of success.

Not all tests are passing when doing 'make check' and
that is not related to this change [1].
While provisioning will invoke 'make check', failures will
be ignored based on this variable: exit_rc_when_failed.

The provisioning does not yet include building rpms+deb
packages.

[1]: https://github.com/ovn-org/ovn/issues/23

Signed-off-by: Flavio Fernandes <flavio@flaviof.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>