Fix testsuite 85 - "ensure one gw controller restart in HA doesn't bounce the master ok".
This testsuite is failing frequently in travis CI and locally when tests are run with "-j5".

The test case deletes the chassis row for chassis 'gw2' and expects that this
doesn't cause the chassisresident port on the master ('gw1') to not bounce as 'gw1'
has higher priority than gw2. But since the ha chassis group has only 2 chassis,
'gw2' can claim the port momemtarily when 'gw2' chassis row is recreated back until
BFD session with 'gw1' is not established.

This patch changes the test assertion approach and makes sure that 'gw1' is the
owner of the chassisresident port eventually.

Signed-off-by: Numan Siddique <numans@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>