northd: Add `is-active` management command
When ovn-northd is connected to clustered OVN DB servers a OVSDB
locking feature is used to ensure only one ovn-northd process is
active at a time.

This patch adds a `is-active` management command that allows the
operator to query whether a ovn-northd process is currently active
or not.

At present this information is only available in the log.

Signed-off-by: Frode Nordahl <frode.nordahl@canonical.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>