ovn-controller: Add command to trigger an I-P full recompute.
Incremental processing tries to minimize the number of times
ovn-controller has to fully reprocess the contents of the southbound
database. However, if a bug in the I-P code causes ovn-controller to
end up in an inconsistent state, we have no easy way to force a full
recalculation of the openflow entries.

This commit adds a new command to ovn-controller, "recompute", which
allows users to force a full recompute of the database. It can be
triggered by the user in the following way:

ovn-appctl -t ovn-controller recompute

Reviewed-by: Daniel Alvarez <dalvarez@redhat.com>
Signed-off-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>