ovn-controller: Run I-P engine even when no SB txn is available.
If the last ovn-controller main loop run started a transaction to the
Southbound DB and the transaction is still in progress, ovn-controller
will not call engine_run(). In case there were changes to the DB,
engine_need_run() would return true which would trigger an immediate
forced recompute in the next processing loop iteration.

However, there are scenarios when updates can be processed incrementally
even if no Southbound DB transaction is available. One example, often
seen in the field, is when the MAC_Binding table is updated. Currently
such an update received while a transaction is still committing would
trigger a forced recompute.

To minimize the number of forced recomputes, we now call
engine_run(false), i.e., try to process updates incrementally without
allowing recompute, also when ovnsb_idl_txn == NULL. This is safe
because ovnsb_idl_txn is not used by change_handlers and run handlers
are not allowed to execute when calling engine_run(false).

To make sure that other users of ovnsb_idl_txn, like pinctrl_run(),
get a chance to run as soon as the transaction is completed, if the
engine has successfully run and ovnsb_idl_txn is NULL we trigger an
immediate wake and a new iteration of the processing loop.

CC: Han Zhou <hzhou@ovn.org>
CC: Numan Siddique <numans@ovn.org>
Signed-off-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>