Add a way to delete QoS by its name or uuid
Currently, qos can only be deleted by indirect way which must designate
switch or more parameters. This patch change the original "qos-del" to
"ls-qos-del", and add a new "qos-del" command. By the new command, you
can delete qos by uuid or name of the qos. It is very import to support
a way to delete a table by direct way. For example, we can associate
the qos table in neutron and OVN by "name" of qos in OVN, so neutron
could find and easily delete the corresponding qos in OVN.

Signed-off-by: Yunxiang Tao <taoyunxiang@cmss.chinamobile.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>