ovn-nbctl.c: Fix lr-route-add/del command with policy considered.
When adding a new route, if the prefix existed but the policy is
different, then they are different routes and the new one should
be added, instead of updating the old one that has different
policy. For example, below commands should add two routes.

ovn-nbctl lr-route-add lr1 10.0.0.0/24 192.168.0.1
ovn-nbctl --policy=src-ip lr-route-add lr1 10.0.0.0/24 192.168.0.2

Without this patch, the second command will fail. It would succeed
with --may-exist, but it would end up with only the second one added.

Similarly, this patch fixes the ambiguity for lr-route-del command
by adding --policy option to specify the policy together with
prefix for matching the route that is supposed to be deleted.
If --policy is not specified, it deletes all routes with the specified
prefix. Without this patch, the command would delete a route randomly
with the given prefix, and there was no way to acurrately specify the
route to be deleted when there are multiple routes with same prefix
but different policy.

Signed-off-by: Han Zhou <hzhou@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>