ovn-northd: Support ECMP routes.
This patch supports ECMP routes for OVN routers. When there are multiple
routes with same prefix and policy, OVN will select one of the routes
based on 5-tuple hash of the packet header with the help of the "select"
action added earlier.

This patch also updates ovn-nbctl so that it allows adding multiple
routes with same prefix and policy when option --ecmp is specified,
and also supports deleting a specific route among ECMP routes
by specifying nexthop and output_port fields.

Signed-off-by: Han Zhou <hzhou@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>