ovn-controller: No bridge for localnet port is not an error
There is a pattern among CMSes to create a `localnet` port
binding without any chassis affiliation.

It is then up to the user to configure chassis with external
connectivity by adding a mapping under the
`external_ids:ovn-bridge-mappings` key in the Open_vSwitch
table of the local OVS instance.  This may be some chassis
or all chassis depending on end user requirements.

At present `ovn-controller` will repeatedly log an error on
chassis without mapping for a localnet port while in reality
it is not an error.

Sined-off-by: Frode Nordahl <frode.nordahl@canonical.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>