Add SCTP support to load balancers.
This allows for load balancers to use SCTP as a supported protocol in
addition to the already-supported UDP and TCP.

With this patch, health checks are not supported for SCTP load
balancers. A test has been added to ensure that this is the case. Health
checks should be added for SCTP load balancers in the near future. When
that's done, the existing test can be updated to ensure that the SCTP
health check works properly.

Signed-off-by: Mark Michelson <mmichels@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>