Rename "VTEP switch" to "ramp switch".
The term "VTEP" is a poor choice because a VTEP is just a VXLAN Tunnel
Endpoint, which is a generic term not specific to OVN.  When we talk
about "VTEP switches" and "VTEP gateways", therefore, it's somewhat
confusing.

This commit introduces the term "ramp", which uses the metaphor of one
of these switches as an on-ramp and an off-ramp to an OVN logical
network, and substitutes it wherever it can be in the documentation.
Bits of the OVN database are harder to change, since changing them
would cause backward compatibility problems, so this commit doesn't try.

It might be a good idea to change some program names (e.g.
ovn-controller-vtep), some distro package names, and so on.  This commit
does not do that.

CC: Ihar Hrachyshka <ihrachys@redhat.com>
Signed-off-by: Ben Pfaff <blp@ovn.org>
Acked-by: Numan Siddique <numans@ovn.org>
Signed-off-by: Numan Siddique <numans@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>