Rename "ovn-controller-vtep" to "ovn-controller-ramp"
The previous patch in this series, introduced the term
"ramp" which  subsituted for the usage of "vtep" in the
documentation and other places. This patch renames the
program name "ovn-controller-vtep" to "ovn-controller-ramp"
and also attemps to rename the usage of term "vtep" to "ramp"
in the code base.

This patch doesn't attemp to rename in the places where it is hard.
This could be addressed as follow up patches.

Signed-off-by: Numan Siddique <numans@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>