ovn-controller: Skip vport bindings done through OVS external_ids:iface-id.
Port bindings of type "virtual" should not have an associated OVS port
in the integration bridge. If this is the case, it's a misconfig and
ovn-controller should ignore it.

If such a situation is detected, ovn-controller will also log a warning
message to inform the user about the wrong configuration.

Reported-at: https://bugzilla.redhat.com/1818844
CC: Numan Siddique <nusiddiq@redhat.com>
Fixes: 054f4c85c413 ("Add a new logical switch port type - 'virtual'")
Signed-off-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>