ovn-nbctl: Create daemon control socket in ovn run dir
ovn-nbctl when run as a daemon is creating the ctl socket in
the ovs rundir. This patch fixes this issue by creating it in
the ovn rundir.

When an ovn service is run with -u option (which specifies the
ctl socket path) and if this path is not absolute, the ovn
ctl socket path is created in the ovs run dir. This patch
also fixes this issue by creating it in the ovn run dir.

Reported-by: Dan Williams <dcbw@redhat.com>
Signed-off-by: Numan Siddique <numans@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>